Add lintian override for bash-term-in-posix-shell wrt. ulimit
This is supported by dash and mksh and these are the only shells I am
aware of which are used in Debian as /bin/sh besides bash.

Not sure why lintian outputs one of the two "&" in the warning as
well, but even the full line works as expected in dash and mksh.